Make all codegen handlers responsible for adjusting the stack
"let" used to restore %sp after its body.  But if its body was a
conditional, it would end in a branch, and the %sp correction would never
be executed.  So pass a "before frame base" and "after frame base" to
all codegens, and make them responsible for correcting %sp, if necessary.
That way, conditionals can put the correction in the right place, before
the branch.  This also naturally folds %sp changes together, which is
nice.