Don't allocate semispace at each GC
By allocating a semispace at each GC, we require the kernel to zero pages
on demand as the fresh pages get faulted in.  Better to allocate the
semispaces on startup and re-use the pages.  We can do this safely now
that we clear all objects on allocation.  This has the added benefit that
we clear memory closer to its use, improving locality.