Allow immediates as general results on ARM
For a better treatment of immediates on ARM, we need to allow
immediates as general results (i.e. values returned from "codegen").
The obstacle to doing this is that ARM can only emit-frame-push
registers, not immediates; immediates need moving into a register
before they are pushed.

Hence emit-frame-spill: It's like emit-frame-push, but accepts a
general operand, and takes a register to be used if necessary to push
that operand.