x86: Refine copy-mem
Previously, we would decide whether to do a forwards copy or a backwards
copy based on whether the destination address was higher or lower than
the source address.  But it only matters which we do if the source and
destination overlap.  And it seems that a backwards REP MOVS is slower
than a forwards one.

So, refine the condition to only do a backwards copy when the destination
address falls within the source region.

Also, move all the logic to determine which kind of copy to do to canned
assembly.  It's actually simpler that way, as well as yielding better
code in a very hot path.