Generate slightly better code for apply
raw-jump-with-arg-space was generating an indirect jump, even though
the label to jump to was statically known.  By sharing the normal
function jump code, we emit a direct jump.  And omit clearing %nargs.

This change also means that we can generate a direct jump for
((lambda (...) ...) ...), but that pattern doesn't occur anywhere.