Kill when-compiling, and don't expand the program twice
Rather than using the when-compiling hack, and expanding the program
twice, we can mark some source files as not to be fed to the interpreter
(they still get expanded, but the forms resulting from expansion are only
added to the program to be compiled).  These source files are marked
with a 'no-interp!' prefix.  It's not pretty, but it's less ugly than
when-compiling, and avoiding repeated expansion yields a nice performance
boost.