Mark self calls as not needing the closure
For self calls, %closure already contains the closure, so there is
no need to evaluate it.  This doesn't affect the generated code currently,
as emit-mov-maybe would squish a redundant mov from %closure to %closure.
But it is more useful if we want to do more sophistication code generation
for calls.