Generate more efficient code for tail calls
The approach to code generation for tail calls was very simple:  The
generated code would push the argument values onto the bottom of the
stack (as if doing a non-tail call), and then copy those values into
their correct locations.  So the code might do a lot of redundant
loads and stores from and to the stack.

We now generate code to evaluate the argument values into their natural
locations (registers or frame slots), and then generate code to store
those values into the correct locations on the stack.  Where arguments
values are already on the stack, we have to shuffle the stack, which
involves traversing the graph of dependencies between stack locations
so that we copy things in a suitable order.