Prefix function code with the arity mismatch check when we can
When a function is only called indirectly, it is good to place
the arity mismatch check in front of the function's code, as we used to
do.  This avoids a branch in the indirect call case, and performance
is important for some indirect calls, such as in the GC type dispatch
table.