x86-64: Support for PIE
x86-64 code generation was sloppy about putting label expressions into
32-bit immediate fields.  This produces R_X86_64_32S relocations in
the ELF files, truncating the result to 32-bits.  This worked because
the linux x86-64 toolchain used to default to a "small" model, which
loads the main executable into the first 2GB of the address space.

Now gcc defaults to PIE, allowing the executable to be loaded at any
address.  This produces errors for R_X86_64_32S relocations.

The solution is to use 'immediate32?' more widely to check whether an
operand is suitable for a 32-bit field.  A few places needed fixing
up appropriately.